requirements.txt
python-crontab
slack_client